Cleanup, work on model view matrix
- Work on fix for ProjectionMatrix
- Add some methods on Matrix classes
- Cleanup & refactoring

+++ Source/Math/BaseMath.cpp
@@ -1,6 +1,6 @@
 #include "BaseMath.h"
 
-double BaseMath::toRadians(double angleDegree)
+float BaseMath::toRadians(float angleDegree)
 {
     return angleDegree / 180.0 * PI;
 }
